add modal input with minimal impact
Implements vi-style modal input.  Each mode has a distinct set of key
shortcuts, plus the ability to decide whether unrecognized keys are
ignored or passed through.

This is a "minimal impact" change in that its default configuration
does not alter the behavior of surf.  A tiny Normal mode which only
accepts the `i' key is given as a demonstration of how to use the
setmode function and ignore unrecognized key events.